requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
mysql-connector-python==8.3.0
